Raise a RuntimeError if asyncio is not available.
Fixes #830.
