+++ setup.py
@@ -5,15 +5,12 @@
 
 
 import os
-from setuptools import setup
-from setuptools.command.test import test as TestCommand
-
 import sys
 
+from setuptools import setup, find_packages
+from setuptools.command.test import test as TestCommand
+
 from gunicorn import __version__
-
-
-ASYNCIO_COMPAT = sys.version_info >= (3, 3)
 
 
 CLASSIFIERS = [
@@ -71,17 +68,6 @@ class PyTest(TestCommand):
 
 REQUIREMENTS = []
 
-py_modules = []
-
-for root, folders, files in os.walk('gunicorn'):
-    for f in files:
-        if f.endswith('.py') and (ASYNCIO_COMPAT or f != 'gaiohttp.py'):
-            full = os.path.join(root, f[:-3])
-            parts = full.split(os.path.sep)
-            modname = '.'.join(parts)
-            py_modules.append(modname)
-
-
 setup(
     name = 'gunicorn',
     version = __version__,
@@ -95,7 +81,7 @@ setup(
 
     classifiers = CLASSIFIERS,
     zip_safe = False,
-    py_modules = py_modules,
+    packages = find_packages(exclude=['examples', 'tests']),
     include_package_data = True,
 
     tests_require = tests_require,
